+++ scripts/rulesCoverage.ts
@@ -1,30 +1,25 @@
 /* eslint-disable @typescript-eslint/no-require-imports */
 import fs from 'fs';
 import path from 'path';
+import rimraf from 'rimraf';
 
 import { NAMESPACE_CONFIG, NAMESPACES } from '../config';
 
+const allRuleMap: Map<string, any> = new Map();
 /** 可用的规则（去除废弃的和 Prettier 的规则） */
 const activeRules: string[] = [];
 const deprecatedRules: string[] = [];
-const prettierRules = [
-  ...Object.keys(require('eslint-config-prettier').rules),
-  ...Object.keys(require('eslint-config-prettier/@typescript-eslint').rules),
-  ...Object.keys(require('eslint-config-prettier/babel').rules),
-  ...Object.keys(require('eslint-config-prettier/flowtype').rules),
-  ...Object.keys(require('eslint-config-prettier/react').rules),
-  ...Object.keys(require('eslint-config-prettier/standard').rules),
-  ...Object.keys(require('eslint-config-prettier/unicorn').rules),
-  ...Object.keys(require('eslint-config-prettier/vue').rules),
-];
+// https://github.com/prettier/eslint-config-prettier/pull/217
+const prettierRules = [...Object.keys(require('eslint-config-prettier').rules), 'vue/quote-props'];
 
 // 填充 deprecatedRules 和 activeRules
 Object.values(NAMESPACE_CONFIG).forEach(({ rulePrefix, pluginName }) => {
   const ruleEntries = pluginName
     ? Object.entries<any>(require(pluginName).rules)
-    : Array.from<any>(require('eslint/lib/rules').entries());
+    : Array.from<any>(require('eslint/use-at-your-own-risk').entries());
   ruleEntries.forEach(([ruleName, ruleValue]) => {
     const fullRuleName = rulePrefix + ruleName;
+    allRuleMap.set(fullRuleName, ruleValue);
     if (ruleValue.meta.deprecated) {
       deprecatedRules.push(fullRuleName);
       return;
@@ -48,15 +43,23 @@ const remainingRules = [...activeRules];
  */
 NAMESPACES.forEach((namespace) => {
   fs.readdirSync(path.resolve(__dirname, '../test', namespace))
-    .filter((ruleName) => fs.lstatSync(path.resolve(__dirname, '../test', namespace, ruleName)).isDirectory())
+    .filter((ruleName) => fs.existsSync(path.resolve(__dirname, '../test', namespace, ruleName, '.eslintrc.js')))
     .forEach((ruleName) => {
       const fullRuleName = NAMESPACE_CONFIG[namespace].rulePrefix + ruleName;
+      const fullRuleDir = getFullRuleDir(fullRuleName);
       if (deprecatedRules.includes(fullRuleName)) {
-        errors.push(`${fullRuleName} 是已废弃的规则，请删除`);
+        const errorMessage = `${fullRuleName} 是已废弃的规则，请删除`;
+        errors.push(errorMessage);
+        errors.push(errorMessage);
+        rimraf.sync(fullRuleDir);
+        console.error(errorMessage);
         return;
       }
       if (prettierRules.includes(fullRuleName)) {
-        errors.push(`${fullRuleName} 是 Prettier 的规则，请删除`);
+        const errorMessage = `${fullRuleName} 是 Prettier 已忽略的规则，会自动删除`;
+        errors.push(errorMessage);
+        rimraf.sync(fullRuleDir);
+        console.error(errorMessage);
         return;
       }
       if (activeRules.includes(fullRuleName)) {
@@ -65,6 +68,19 @@ NAMESPACES.forEach((namespace) => {
       }
     });
 });
+
+export function getFullRuleDir(fullRuleName: string) {
+  if (!fullRuleName.includes('/')) {
+    return path.resolve(__dirname, '../test/base', fullRuleName);
+  }
+  const [prefix, rule] = fullRuleName.split('/');
+  for (const [namespace, config] of Object.entries(NAMESPACE_CONFIG)) {
+    if (config.rulePrefix === `${prefix}/`) {
+      return path.resolve(__dirname, '../test', namespace, rule);
+    }
+  }
+  return '';
+}
 
 function getDocsUrlFromRuleName(ruleName: string) {
   for (const namespaceConfig of Object.values(NAMESPACE_CONFIG).reverse()) {
@@ -76,13 +92,30 @@ function getDocsUrlFromRuleName(ruleName: string) {
 }
 
 if (remainingRules.length > 0) {
-  errors.push(`
-  以下规则需要补充：${remainingRules
-    .map((ruleName) => `\n  - ${ruleName} ${getDocsUrlFromRuleName(ruleName)}`)
-    .join('')}
-  `);
+  remainingRules.forEach((fullRuleName) => {
+    const fullRuleDir = getFullRuleDir(fullRuleName);
+    const ruleValue = allRuleMap.get(fullRuleName);
+    if (!fs.existsSync(fullRuleDir)) {
+      fs.mkdirSync(fullRuleDir);
+    }
+    fs.writeFileSync(
+      path.resolve(fullRuleDir, '.eslintrc.js'),
+      `module.exports = {
+  rules: {
+    /**
+     * ${ruleValue.meta.docs.description}
+     */
+    '${fullRuleName}': 'off',
+  },
+};
+`,
+    );
+    const errorMessage = `${fullRuleName} 是一条自动添加的新规则, , ${getDocsUrlFromRuleName(fullRuleName)}`;
+    errors.push(errorMessage);
+    console.error(errorMessage);
+  });
 }
 
 if (errors.length > 0) {
-  throw new Error(errors.map((error) => `\n- ${error}`).join(''));
+  process.exit(1);
 }