+++ test/vue/valid-v-bind-sync/good.vue
@@ -1,7 +1,0 @@
-<template>
-  <MyComponent :aaa.sync="foo" />
-
-  <div v-for="todo in todos">
-    <MyComponent :aaa.sync="todo.name" />
-  </div>
-</template>