+++ test/vue3/component-definition-name-casing/bad.vue
@@ -1,5 +1,0 @@
-<script>
-export default {
-  name: 'foo-bar',
-};
-</script>