+++ test/vue3/no-empty-pattern/.eslintrc.js
@@ -1,7 +1,7 @@
 module.exports = {
   rules: {
     /**
-     * 禁止解构赋值中出现空 {} 或 []
+     * 禁止在 template 里解构赋值中出现空 {} 或 []
      */
     'vue/no-empty-pattern': 'error',
   },