+++ test/vue3/no-parsing-error/.eslintrc.js
@@ -1,7 +1,7 @@
 module.exports = {
   rules: {
     /**
-     * 禁止出现语法错误
+     * 禁止在 template 里出现语法错误
      */
     'vue/no-parsing-error': 'error',
   },