+++ test/vue3/no-restricted-syntax/.eslintrc.js
@@ -1,7 +1,7 @@
 module.exports = {
   rules: {
     /**
-     * 禁止使用指定的语法
+     * 禁止在 template 里使用指定的语法
      */
     'vue/no-restricted-syntax': 'off',
   },